requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
async-timeout==4.0.3
attrs==21.2.0
Babel==2.8.0
blinker==1.8.2
certifi==2020.6.20
chardet==4.0.0
click==8.1.7
cloud-init==24.1.3
colorama==0.4.4
configobj==5.0.6
cryptography==3.4.8
dbus-python==1.2.18
distro==1.7.0
distro-info==1.1+ubuntu0.2
Flask==3.0.3
frozenlist==1.4.1
httplib2==0.20.2
idna==3.3
importlib-metadata==4.6.4
influxdb==5.3.2
itsdangerous==2.2.0
jeepney==0.7.1
Jinja2==3.1.4
jsonpatch==1.32
jsonpointer==2.0
jsonschema==3.2.0
keyring==23.5.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
load-dotenv==0.1.0
MarkupSafe==2.1.5
more-itertools==8.10.0
msgpack==1.0.8
multidict==6.0.5
netifaces==0.11.0
numpy==1.26.4
oauthlib==3.2.0
psutil==5.9.8
PyGObject==3.42.1
PyJWT==2.3.0
PyMySQL==1.1.1
pyparsing==2.4.7
pyrsistent==0.18.1
pyserial==3.5
python-apt==2.4.0+ubuntu3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2022.1
PyYAML==5.4.1
requests==2.25.1
SecretStorage==3.3.1
six==1.16.0
ssh-import-id==5.11
systemd-python==234
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.26.5
wadllib==1.3.6
Werkzeug==3.0.3
xkit==0.0.0
yarl==1.9.4
zipp==1.0.0
